Re-enable S3 interrupt handler.
S3 can now be set manually via LED0.  When this happens, the S3
interrupt handler is called, and S3 processing goes on as if the
S3 signal was generated automatically by the hardware (it still can
be generated by the glance hardware as a fail safe when the main
voltage falls below the switchover threshold).